[SPARK-54657][PS] Refactor pyspark.sql.pandas.serializers for readability/reuse
